% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRT-package.R
\docType{package}
\name{cIRT-package}
\alias{cIRT-package}
\alias{_PACKAGE}
\title{cIRT: Choice Item Response Theory}
\description{
Jointly model the accuracy of cognitive responses and item choices
    within a Bayesian hierarchical framework as described by Culpepper and
    Balamuta (2015) <doi:10.1007/s11336-015-9484-7>. In addition, the package
    contains the datasets used within the analysis of the paper.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tmsalab.github.io/cIRT}
  \item \url{https://github.com/tmsalab/cIRT}
  \item Report bugs at \url{https://github.com/tmsalab/cIRT/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{balamut2@illinois.edu} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Steven Andrew Culpepper \email{sculpepp@illinois.edu} (\href{https://orcid.org/0000-0003-4226-6176}{ORCID}) [copyright holder]
}

}
