% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cIRT-package.R
\docType{package}
\name{cIRT_package}
\alias{cIRT_package}
\alias{cIRT_package-package}
\alias{cirt}
\title{Item Response Theory Models with Guided Choice}
\description{
Item Response Theory Models with Guided Choice
}
\details{
\tabular{ll}{
Package: \tab cIRT\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2015-12-05\cr
License: \tab GPL-3\cr
}
}
\author{
Authors@R: c(person("Steve", "Culpepper", role = c("aut", "cph"), email =
"sculpepp@illinois.edu"), 
person("James", "Balamuta", role = c("aut","cre", "ctb", "com","cph"),email = "balamut2@illinois.edu"))

James Balamuta balamut2@illinois.edu
}
\seealso{
\code{\link{cIRT}}, \code{\link{TwoPLChoicemcmc}}, \code{\link{probitHLM}}, \code{\link{direct_sum}},  \code{\link{center_matrix}},
          \code{\link{rmvnorm}}, \code{\link{rwishart}}, \code{\link{riwishart}},
          \code{\link{payout_matrix}}, \code{\link{survey_data}},
          \code{\link{trial_matrix}}, and \code{\link{choice_matrix}}
}
\keyword{package}

