% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payout_matrix.R
\name{payout_matrix}
\alias{payout_matrix}
\title{Payout Matrix Data}
\format{A data frame with 252 observations on the following 4 variables.
\describe{
  \item{\code{Participant}}{Subject ID}
  \item{\code{cum_sum}}{Sum of all payouts}
  \item{\code{num_correct_choices}}{Total number of correct choices (/15)}
  \item{\code{num_correct_trials}}{Total number of correct trials (/30)}
}}
\source{
Choice38 Experiment at UIUC during Spring 2014 - Fall 2014
}
\usage{
data(payout_matrix)
}
\description{
This data set contains the payout information for each subject.
}
\author{
Steven Culpepper and James Balamuta
}
