% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Survey Data}
\format{
A data frame with 102 observations on the following 2 variables.
\describe{
\item{\code{id}}{Subject's Assigned Research ID}
\item{\code{sex}}{Subject's sex:
\itemize{
\item Male
\item Female
}
}
}
}
\source{
Choice38 Experiment at UIUC during Spring 2014 - Fall 2014
}
\usage{
survey_data
}
\description{
This data set contains the subject's responses survey questions administered
using Choice38.
}
\author{
Steven Andrew Culpepper and James Joseph Balamuta
}
\keyword{datasets}
