% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{payout_matrix}
\alias{payout_matrix}
\title{Payout Matrix Data}
\format{
A data frame with 252 observations on the following 4 variables.
\describe{
\item{\code{Participant}}{Subject ID}
\item{\code{cum_sum}}{Sum of all payouts}
\item{\code{num_correct_choices}}{Total number of correct choices (out of 15)}
\item{\code{num_correct_trials}}{Total number of correct trials (out of 30)}
}
}
\source{
Choice38 Experiment at UIUC during Spring 2014 - Fall 2014
}
\usage{
payout_matrix
}
\description{
This data set contains the payout information for each subject.
}
\author{
Steven Andrew Culpepper and James Joseph Balamuta
}
\keyword{datasets}
